requirements.txt
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.4
ImageHash==4.3.1
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
scipy==1.10.1
six==1.16.0
